+++ common/i18n/src/test/java/is/codion/common/i18n/MessagesTest.java
@@ -28,5 +28,8 @@ public class MessagesTest {
     Messages.clearMnemonic();
     Messages.advanced();
     Messages.find();
+    Messages.refresh();
+    Messages.refreshMnemonic();
+    Messages.refreshTip();
   }
 }