+++ framework/i18n/src/test/java/is/codion/framework/i18n/FrameworkMessagesTest.java
@@ -28,9 +28,6 @@ public class FrameworkMessagesTest {
     FrameworkMessages.deleteMnemonic();
     FrameworkMessages.deleteCurrentTip();
     FrameworkMessages.deleteSelectedTip();
-    FrameworkMessages.refresh();
-    FrameworkMessages.refreshMnemonic();
-    FrameworkMessages.refreshTip();
     FrameworkMessages.dependencies();
     FrameworkMessages.dependenciesTip();
     FrameworkMessages.add();